edit: gql config
