+++ package.json
@@ -13,7 +13,8 @@
 		"test:unit": "vitest",
 		"lint": "prettier --plugin-search-dir . --check . && eslint .",
 		"format": "prettier --plugin-search-dir . --write .",
-		"codegen": "graphql-codegen"
+		"codegen": "graphql-codegen",
+		"codegen:watch": "graphql-codegen --watch"
 	},
 	"devDependencies": {
 		"@fontsource/fira-mono": "^4.5.10",
@@ -38,7 +39,7 @@
 		"prettier-plugin-svelte": "^2.8.1",
 		"prettier-plugin-tailwindcss": "^0.3.0",
 		"sass": "^1.62.0",
-		"svelte": "^3.54.0",
+		"svelte": "^3.59.2",
 		"svelte-breakpoints": "^0.2.0",
 		"svelte-check": "^3.0.1",
 		"svelte-preprocess": "^5.0.3",
@@ -48,7 +49,8 @@
 		"typescript": "^5.1.6",
 		"vite": "^4.2.0",
 		"vite-plugin-node-polyfills": "^0.8.1",
-		"vitest": "^0.25.3"
+		"vitest": "^0.25.3",
+		"svelte2tsx": "^0.6.16"
 	},
 	"type": "module",
 	"dependencies": {
