+++ src/gql/index.ts
